Refactor: mock 폴더를 mocks 폴더로 변경
Resolves:
Ref:
